requirements.txt
Django==5.2.6
google-generativeai
python-dotenv